dto funcionando

+++ api/PopCulture.Api/Models/LocationDto.cs
@@ -2,9 +2,9 @@ namespace PopCulture.Api.Models
 {
     public class LocationDto
     {
-        public int Id { get; set; }
-        public string Name { get; set; } = string.Empty;
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public long   Id        { get; set; }       // era int, virou long
+        public string Name      { get; set; } = string.Empty;
+        public double Latitude  { get; set; }       // era decimal, virou double
+        public double Longitude { get; set; }       // era decimal, virou double
     }
 }
